added database creation to the flask_app directory
