--- database/configs/__init__.py
@@ -1,1 +1,0 @@
-from .siloing_data_config import *
