+++ flask_app/dummy_data/siloing_data_config.py
@@ -2,5 +2,5 @@ def dummy_siloing_data():
     return [
         {'token_id': '1234123412341234', 'platform': 'Youtube', 'time': 123},
         {'token_id': '9123491234', 'platform': 'Facebook', 'time': 1234}
-        ]
+    ]
